pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "wallet"
version = "2.3.0-alpha.0"
description = "Personal finance service"
authors = ["Kirill Sumorokov <sumorokov.k@gmail.com>"]
license = "MIT"
packages = [
    { include = "wallet", from = "src" }
]

[tool.poetry.dependencies]
aiohttp-metrics = {git = "https://github.com/clayman083/aiohttp-metrics.git", rev = "v0.1.2"}
aiohttp-micro = { path = "../aiohttp_micro/" }
# aiohttp-micro = {git = "https://github.com/clayman083/aiohttp-micro.git", rev = "v0.2.4"}
asyncpg = "^0.20.1"
passport = { path = "../passport/" }
# passport = {git = "https://github.com/clayman083/passport.git", rev = "v2.5.1"}
python = "^3.8"
cerberus = "^1.3.2"
pendulum = "^2.1.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
coverage = "*"
faker = "*"
flake8 = "*"
flake8-bugbear = "*"
flake8-builtins-unleashed = "*"
flake8-comprehensions = "*"
flake8-import-order = "*"
flake8-print = "*"
flake8-pytest = "*"
mypy = "*"
pytest = "*"
pytest-aiohttp = "*"
pytest-postgres = "*"
rope = "*"
ipython = "^7.12.0"
